[리팩터링] 로또즈 팩터리 통합
